Delete `maud_htmlescape` package
Replace it with a symlink. Cargo will automatically resolve this to a
normal file on publish.
