+++ Cargo.toml
@@ -1,6 +1,5 @@
 [workspace]
 members = [
-    "maud_htmlescape",
     "maud_macros",
     "maud",
 ]
