+++ maud_macros/Cargo.toml
@@ -12,7 +12,6 @@ edition = "2021"
 
 [dependencies]
 syn = "1.0.8"
-maud_htmlescape = { version = "0.17.0", path = "../maud_htmlescape" }
 quote = "1.0.7"
 proc-macro2 = "1.0.23"
 proc-macro-error = "1.0.0"
