+++ maud_macros/src/escape.rs
@@ -1,0 +1,1 @@
+../../maud/src/escape.rs
